+++ frappe/desk/form/utils.py
@@ -148,4 +148,4 @@ def _sort_field_fallback(doctype: str, fieldname: str):
 
 
 def get_pdf_link(doctype, docname, print_format="Standard", no_letterhead=0):
-	return f"/api/method/frappe.utils.print_format.download_pdf?doctype={doctype}&name={docname}&format={print_format}&no_letterhead={no_letterhead}"
+	return f"/api/method/frappe.printing.api.download_pdf?doctype={doctype}&name={docname}&format={print_format}&no_letterhead={no_letterhead}"
