+++ frappe/public/js/frappe/microtemplate.js
@@ -209,7 +209,7 @@ frappe.render_pdf = function (html, opts = {}) {
 	formData.append("blob", blob);
 
 	var xhr = new XMLHttpRequest();
-	xhr.open("POST", "/api/method/frappe.utils.print_format.report_to_pdf");
+	xhr.open("POST", "/api/method/frappe.printing.api.report_to_pdf");
 	xhr.setRequestHeader("X-Frappe-CSRF-Token", frappe.csrf_token);
 	xhr.responseType = "arraybuffer";
